stm32 interface driver update, do not use stm32_GPIO_INPD or stm32_GPIO_INPU in driver

+++ vsf/interfaces/cpu/stm32/spi/STM32_SPI.c
@@ -425,7 +425,7 @@ vsf_err_t stm32_spi_config(uint8_t index, uint32_t kHz, uint8_t mode)
 				#if SPI00_NSS_ENABLE
 				GPIOA->CRL = (GPIOA->CRL & ~(0x0F << (4 * 4))) | 
 								(uint32_t)stm32_GPIO_INP << (4 * 4);
-				GPIOA->BSRR = 1 << 4;
+				GPIOA->BSRR = (uint32_t)1 << 4;
 				#endif
 			}
 			break;
@@ -465,7 +465,7 @@ vsf_err_t stm32_spi_config(uint8_t index, uint32_t kHz, uint8_t mode)
 				#if SPI10_NSS_ENABLE
 				GPIOA->CRH = (GPIOA->CRH & ~(0x0F << ((15 - 8) * 4))) | 
 								(uint32_t)stm32_GPIO_INP << ((15 - 8) * 4);
-				GPIOA->BSRR = 1 << 15;
+				GPIOA->BSRR = (uint32_t)1 << 15;
 				#endif
 			}
 			break;
@@ -514,7 +514,7 @@ vsf_err_t stm32_spi_config(uint8_t index, uint32_t kHz, uint8_t mode)
 				#if SPI01_NSS_ENABLE
 				GPIOB->CRH = (GPIOB->CRH & ~(0x0F << ((12 - 8) * 4))) | 
 								(uint32_t)stm32_GPIO_INP << ((12 - 8) * 4);
-				GPIOB->BSRR = 1 << 12;
+				GPIOB->BSRR = (uint32_t)1 << 12;
 				#endif
 			}
 			break;
@@ -562,7 +562,7 @@ vsf_err_t stm32_spi_config(uint8_t index, uint32_t kHz, uint8_t mode)
 				#if SPI02_NSS_ENABLE
 				GPIOA->CRH = (GPIOA->CRH & ~(0x0F << ((15 - 8) * 4))) | 
 								(uint32_t)stm32_GPIO_INP << ((15 - 8) * 4);
-				GPIOA->BSRR = 1 << 15;
+				GPIOA->BSRR = (uint32_t)1 << 15;
 				#endif
 			}
 			break;
